main cleanup
